feat(signal): modularize server
