+++ .github/workflows/docker-grpc.yaml
@@ -17,7 +17,7 @@ jobs:
       - uses: actions/checkout@v2
 
       - name: build
-        run: docker build --tag pionwebrtc/ion-avp:latest-grpc -f cmd/server/grpc/Dockerfile .
+        run: docker build --tag pionwebrtc/ion-avp:latest-grpc -f cmd/signal/grpc/Dockerfile .
 
       - name: login
         if: github.event_name == 'release'
