+++ .vscode/launch.json
@@ -1,0 +1,47 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "webm-server",
+            "type": "go",
+            "request": "launch",
+            "mode": "auto",
+            "program": "${workspaceFolder}/examples/save-to-webm/server/main.go",
+            "env": {},
+            "args": ["-c", "config.toml"],
+            "cwd": "${workspaceFolder}"
+        },
+        {
+            "name": "webm-client",
+            "type": "go",
+            "request": "launch",
+            "mode": "auto",
+            "program": "${workspaceFolder}/examples/save-to-webm/client/main.go",
+            "env": {},
+            "cwd": "${workspaceFolder}"
+        },
+        {
+            "name": "mixing-server",
+            "type": "go",
+            "request": "launch",
+            "mode": "auto",
+            "program": "${workspaceFolder}/examples/mixing/server/main.go",
+            "env": {},
+            "args": ["-c", "config.toml"],
+            "cwd": "${workspaceFolder}/examples/mixing/server"
+        },
+        {
+            "name": "mixing-client",
+            "type": "go",
+            "request": "launch",
+            "mode": "auto",
+            "program": "${workspaceFolder}/examples/mixing/client/main.go",
+            "env": {},
+            "args": ["room1", "abc", "55ca4387-6916-466d-81cd-753482c6f996"],
+            "cwd": "${workspaceFolder}"
+        }
+    ]
+}
